+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Under development
 
+- Options
+    - Added
+        - Scope
+            - `--scope-exclude-extensions` -- CSV of file extensions to exclude.
 - `URI` -- Optimized and re-written to completely bypass Ruby's `URI` lib.
 - `HTTP`
     - `ProxyServer` -- Replaced the previous `WEBrick`-based one with a custom